Refactor the AWS External Device Service (#2185)
* #2110 Refactor AWS External Device

- Remove AWS Exterrnal Device
- Calls to AWS APIs in the respective services
- Add methods to mutualize in IExternalDeviceService

* #2110 Add missing TU

* #2110 Add AWSExternalDeviceService TU

+++ src/IoTHub.Portal.Infrastructure/Services/AWS/AWSDeviceService.cs
@@ -10,17 +10,21 @@ namespace IoTHub.Portal.Infrastructure.Services.AWS
     using IoTHub.Portal.Domain.Repositories;
     using IoTHub.Portal.Domain;
     using Amazon.IoT.Model;
-    using IoTHub.Portal.Application.Services.AWS;
     using Amazon.IotData.Model;
     using IoTHub.Portal.Application.Managers;
     using Infrastructure;
     using Microsoft.Extensions.Logging;
+    using IoTHub.Portal.Domain.Exceptions;
+    using Amazon.IoT;
+    using Amazon.IotData;
+    using ResourceNotFoundException = Domain.Exceptions.ResourceNotFoundException;
 
     public class AWSDeviceService : DeviceService
     {
         private readonly IMapper mapper;
         private readonly IDeviceRepository deviceRepository;
-        private readonly IAWSExternalDeviceService externalDevicesService;
+        private readonly IAmazonIoT amazonIoTClient;
+        private readonly IAmazonIotData amazonIotDataClient;
 
         public AWSDeviceService(PortalDbContext portalDbContext,
             IMapper mapper,
@@ -30,26 +34,34 @@ namespace IoTHub.Portal.Infrastructure.Services.AWS
             ILabelRepository labelRepository,
             IDeviceTagService deviceTagService,
             IDeviceModelImageManager deviceModelImageManager,
-            IAWSExternalDeviceService externalDevicesService,
+            IAmazonIoT amazonIoTClient,
+            IAmazonIotData amazonIotDataClient,
             IExternalDeviceService externalDeviceService,
             ILogger<AWSDeviceService> logger)
             : base(mapper, unitOfWork, deviceRepository, deviceTagValueRepository, labelRepository, externalDeviceService, deviceTagService, deviceModelImageManager, null!, portalDbContext, logger)
         {
             this.mapper = mapper;
             this.deviceRepository = deviceRepository;
-            this.externalDevicesService = externalDevicesService;
+            this.amazonIoTClient = amazonIoTClient;
+            this.amazonIotDataClient = amazonIotDataClient;
         }
 
         public override async Task<DeviceDetails> CreateDevice(DeviceDetails device)
         {
             //Create Thing
-            var createThingRequest = this.mapper.Map<CreateThingRequest>(device);
-            var response = await this.externalDevicesService.CreateDevice(createThingRequest);
-            device.DeviceID = response.ThingId;
+            var thingResponse = await this.amazonIoTClient.CreateThingAsync(this.mapper.Map<CreateThingRequest>(device));
+            if (thingResponse.HttpStatusCode != System.Net.HttpStatusCode.OK)
+            {
+                throw new InternalServerErrorException($"Unable to create the thing with device name : {device.DeviceName} due to an error in the Amazon IoT API : {thingResponse.HttpStatusCode}");
+            }
+            device.DeviceID = thingResponse.ThingId;
 
             //Create Thing Shadow
-            var shadowRequest = this.mapper.Map<UpdateThingShadowRequest>(device);
-            _ = await this.externalDevicesService.UpdateDeviceShadow(shadowRequest);
+            var shadowResponse = await this.amazonIotDataClient.UpdateThingShadowAsync(this.mapper.Map<UpdateThingShadowRequest>(device));
+            if (shadowResponse.HttpStatusCode != System.Net.HttpStatusCode.OK)
+            {
+                throw new InternalServerErrorException($"Unable to create/update the thing shadow with device name : {device.DeviceName} due to an error in the Amazon IoT API : {shadowResponse.HttpStatusCode}");
+            }
 
             //Create Thing in DB
             return await CreateDeviceInDatabase(device);
@@ -58,8 +70,11 @@ namespace IoTHub.Portal.Infrastructure.Services.AWS
         public override async Task<DeviceDetails> UpdateDevice(DeviceDetails device)
         {
             //Update Thing
-            var updateThingRequest = this.mapper.Map<UpdateThingRequest>(device);
-            _ = await this.externalDevicesService.UpdateDevice(updateThingRequest);
+            var thingResponse = await this.amazonIoTClient.UpdateThingAsync(this.mapper.Map<UpdateThingRequest>(device));
+            if (thingResponse.HttpStatusCode != System.Net.HttpStatusCode.OK)
+            {
+                throw new InternalServerErrorException($"Unable to update the thing with device name : {device.DeviceName} due to an error in the Amazon IoT API : {thingResponse.HttpStatusCode}");
+            }
 
             //Update Thing in DB
             return await UpdateDeviceInDatabase(device);
@@ -67,10 +82,46 @@ namespace IoTHub.Portal.Infrastructure.Services.AWS
 
         public override async Task DeleteDevice(string deviceId)
         {
-            //Delete Thing
+            //Get device in DB
             var device = await deviceRepository.GetByIdAsync(deviceId);
-            var deleteThingRequest = this.mapper.Map<DeleteThingRequest>(device);
-            _ = await this.externalDevicesService.DeleteDevice(deleteThingRequest);
+
+            if (device == null)
+            {
+                throw new ResourceNotFoundException($"The device with id {deviceId} doesn't exist");
+            }
+
+            //Retrieve all thing principals and detach it before deleting the thing
+            var principals = await this.amazonIoTClient.ListThingPrincipalsAsync(new ListThingPrincipalsRequest
+            {
+                NextToken = string.Empty,
+                ThingName = device.Name
+            });
+
+            if (principals.HttpStatusCode != System.Net.HttpStatusCode.OK)
+            {
+                throw new InternalServerErrorException($"Unable to retreive Thing {device.Name} principals due to an error in the Amazon IoT API : {principals.HttpStatusCode}");
+            }
+
+            foreach (var principal in principals.Principals)
+            {
+                var detachPrincipal = await this.amazonIoTClient.DetachThingPrincipalAsync(new DetachThingPrincipalRequest
+                {
+                    Principal = principal,
+                    ThingName = device.Name
+                });
+
+                if (detachPrincipal.HttpStatusCode != System.Net.HttpStatusCode.OK)
+                {
+                    throw new InternalServerErrorException($"Unable to detach Thing {device.Name} principal due to an error in the Amazon IoT API : {detachPrincipal.HttpStatusCode}");
+                }
+            }
+
+            //Delete the thing type after detaching the principal
+            var deleteResponse = await this.amazonIoTClient.DeleteThingAsync(this.mapper.Map<DeleteThingRequest>(device));
+            if (deleteResponse.HttpStatusCode != System.Net.HttpStatusCode.OK)
+            {
+                throw new InternalServerErrorException($"Unable to delete the thing with device name : {device.Name} due to an error in the Amazon IoT API : {deleteResponse.HttpStatusCode}");
+            }
 
             //Delete Thing in DB
             await DeleteDeviceInDatabase(deviceId);
